1 file for filter logic
